pyproject.toml
[tool.poetry]
name = "learnify-backend"
version = "0.1.0"
description = ""
authors = ["Yakser <iaksanov2044@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django-taggit = "^4.0.0"
django-filter = "^23.3"
django-debug-toolbar = "^4.2.0"
django-cors-headers = "^4.2.0"
drf-yasg = "^1.21.7"
flake8 = "^6.1.0"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.0"
django-simple-history = "^3.4.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
